refactor: formalize Detector interface
